refactor: rename constant pkg -> const pkg
